Fix PlotCommitList to set lanes on child-less commits
In PlotCommitList.enter() commits are positioned on lanes for visual
presentation. This implementation was buggy: commits without
children (often the starting points for the RevWalk) are not positioned
on separate lanes.

The problem was that when handling commits with multiple children (thats
where branches fork out) it was not handled that some of the children
may not have been positioned on a lane yet. I fixed that and added a number
of tests which specifically test the layout of commits on lanes.

Bug: 300282
Bug: 320263
Change-Id: I9bc05006ca53bede45cda95ed693d8a00f00afd9
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>