Improve performance of IndexDiff.diff
The IndexDiff.diff() method caused sha1-computation of every file
in the working tree. The method was doing a TreeWalk with ANY_DIFF
filter and if a WorkingTreeIterator is involved this caused the
sha1 computation on every file. But if the treewalk also contains
a DirCacheIterator we can call
WorkingTreeIterator.isModified(DirCacheIterator ...). This method
can tell whether the file is dirty by looking at metadata
(lastmodified,...) first which is much faster than computing the
sha1.

This fix will improve the performance of EGit's commit dialog
significantly.

Change-Id: I99d66d2575906e92f781ea37131a4b35d84211eb
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>
Signed-off-by: Philipp Thun <philipp.thun@sap.com>