Add client side support for authenticating clients by certificates
When pushing to a git server over https it should be possible to authenticate
the client by client certificates. Native git allows this by the configuration
parameters http.sslKey/http.sslCert which are added to jgit with this commit.
Also support for http.sslCAInfo is added which allows checking server
certificates during https communication.

Unit tests for this functionality are included in a subsequent commit which
brings the server side for authentication by client-certificates.

Change-Id: I0674175083e2337cfbb3d3efa26cb4a9a0181123
Signed-off-by: Philipp Thun <philipp.thun@sap.com>
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>