Fix checkout of pathes
The checkout command was producing a inconsistent state of the index
which even confuses native git. Then content sha1 of the touched index
entries was updated, but the lenght and the filemode was not changed.
Later in coding the index entries got automatically corrected (throught
Dircache.checkoutEntry()) but the correction was after persisting the
index to disk. So, the correction was lost.

Change-Id: Ifc6d806fbf96f53c94d9ded0befcc932d943aa04
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>