Fix positioning of lanes in PlotCommitList
There was a severe bug in PlotCommitList.enter() which
caused that we sometimes draw passing lanes through a commit. E.g. in
EGit History view it looked like a certain commit was located on a
certain lane and had certain parents/children although it was actually
located on it's own lane and therefore had different parents/children.
This bug was hard to reproduce because as soon as the topology changed
(e.g. the wrongly positioned commit is merged into the other) the bug
disappeared and the commit was positioned correctly.

In detail the problem was the following:
For PlotCommits which have no children (e.g. a commit on the tip of a
branch) we delay the decision on which lane they belong. Every
PlotCommit who enters the list also checks whether he has such
unpositioned children and finds lanes for them (additionally to finding
a lane for himself). When deciding on the lane for the children we used
the freePositons collection which reflects the current state. But: There
may be lane positions which are currently free but which are not free
for the child. Passing lanes of the child may occupy this position.
These passing lanes are closed between the child and current commit and
that's why for the current commit that lane position is free.

This fix adds a special method to find lanes for children. There we
explicitly check for free lane positions by iterating over all passing
lanes.

Change-Id: I357875a166efb2a126d37ef8413f8ef8c5bba607
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>